+++ RealFarm/src/com/commonsensenet/realfarm/Global.java
@@ -6,7 +6,7 @@ public class Global {
 
 	public static boolean cameraFlag;
 	public static boolean enableAudio = true;
-	public static long plotId = -1;
+	public static int plotId = -1;
 	public static String plotImagePath;
 	public static Bitmap rotated;
 	public static Class<?> selectedAction = null;
